[LOOP-5707] Migrate Presets to LoopKit
