--- Loop/Views/Presets/Components/PresetCard.swift
@@ -1,137 +1,0 @@
-//
-//  PresetCard.swift
-//  Loop
-//
-//  Created by Cameron Ingham on 10/24/24.
-//  Copyright © 2024 LoopKit Authors. All rights reserved.
-//
-
-import LoopAlgorithm
-import LoopKitUI
-import SwiftUI
-import LoopKit
-import LoopCore
-
-struct PresetCard: View {
-    
-    @Environment(\.colorPalette) private var colorPalette
-    @Environment(\.isEnabled) private var isEnabled
-    @Environment(\.temporaryPresetsManager) private var temporaryPresetsManager
-    @EnvironmentObject var displayGlucosePreference: DisplayGlucosePreference
-    
-    let presetId: String
-    let icon: PresetSymbol?
-    let presetName: String
-    let duration: PresetDuration
-    let insulinMultiplier: Double?
-    let correctionRange: ClosedRange<LoopQuantity>?
-    let guardrail: Guardrail<LoopQuantity>?
-    let expectedEndTime: PresetExpectedEndTime?
-    let isScheduled: Bool
-    let activityPresetIsModified: Bool?
-
-    var presetTitle: some View {
-        HStack(spacing: 6) {
-            if let icon, !icon.isEmpty {
-                PresetSymbolView(icon)
-            }
-
-            Text(presetName)
-                .fontWeight(.semibold)
-                .accessibilityIdentifier("text_Preset\(presetName)")
-            
-            if activityPresetIsModified == false {
-                Text(Image(systemName: "checkmark.seal.fill"))
-                    .font(.subheadline)
-                    .foregroundStyle(Color.accentColor)
-            }
-        }
-    }
-
-    var reminderIcon: some View {
-        Text(Image(systemName: "alarm"))
-            .font(.footnote)
-            .foregroundColor(.carbs)
-            .accessibilityLabel(Text("Scheduled reminder"))
-    }
-
-    var presetDuration: some View {
-        Group { Text(Image(systemName: "timer")) + Text(" \(duration.localizedTitle)") }
-            .font(.footnote)
-            .foregroundColor(.secondary)
-            .accessibilityLabel(Text(duration.accessibilityLabel))
-    }
-    
-    var body: some View {
-        VStack(alignment: .leading, spacing: 10) {
-            ViewThatFits(in: .horizontal) {
-                HStack {
-                    VStack(alignment: .leading) {
-                        if let expectedEndTime {
-                            HStack(spacing: 8) {
-                                Text(Image(systemName: "timer"))
-                                +
-                                Text(" \(expectedEndTime.localizedTitle)")
-                                    .accessibilityLabel(Text(expectedEndTime.accessibilityLabel))
-                            }
-                            .font(.footnote)
-                            .foregroundColor(.white)
-                            .padding(.horizontal, 6)
-                            .padding(.vertical, 5)
-                            .background(Color.presets)
-                            .cornerRadius(8)
-                        }
-                        presetTitle
-                    }
-
-                    Spacer()
-
-                    if expectedEndTime == nil {
-                        presetDuration
-                        if isScheduled {
-                            reminderIcon
-                        }
-                    }
-                }
-                
-                VStack(alignment: .leading, spacing: 10) {
-                    presetTitle
-                    
-                    presetDuration
-                }
-            }
-            
-            Divider()
-                .padding(.horizontal, -10)
-            
-            PresetStatsView(
-                insulinMultiplier: insulinMultiplier,
-                correctionRange: correctionRange,
-                guardrail: guardrail,
-                therapySettingsImpactDisplayState: .hide,
-                isScheduled: isScheduled && expectedEndTime != nil,
-                isActive: temporaryPresetsManager.activePreset?.id == presetId
-            )
-        }
-        .padding(10)
-        .background(RoundedRectangle(cornerRadius: 8)
-            .fill(Color(UIColor.tertiarySystemBackground))
-            .stroke(Color(UIColor.secondarySystemBackground), lineWidth: 1)
-            .frame(maxWidth: .infinity))
-        .opacity(isEnabled ? 1 : 0.6)
-    }
-}
-
-extension Color {
-    init(presetSymbolTint: PresetSymbol.SymbolTint?, palette: LoopUIColorPalette) {
-        guard let presetSymbolTint else {
-            self = .primary
-            return
-        }
-        
-        switch presetSymbolTint {
-        case .preMeal:
-            self = palette.carbTintColor
-        }
-    }
-}
