+++ Loop/Views/Presets/Components/PresetDetentView.swift
@@ -176,7 +176,8 @@ struct PresetDetentView: View {
                     guardrail: settingsManager.correctionRangeGuardrailForPreset(preset),
                     therapySettingsImpactDisplayState: operation == .end ? .show(settingsImpact) : .hide,
                     isScheduled: false,
-                    isActive: temporaryPresetsManager.activePreset?.id == preset.id
+                    isActive: temporaryPresetsManager.activePreset?.id == preset.id,
+                    effectiveCorrectionRange: temporaryPresetsManager.effectiveCorrectionRange
                 )
                 .padding(.horizontal)
