+++ Loop/Views/Presets/Training/Components/PlayMediaButton.swift
