+++ Loop/Views/Presets/Training/Components/TherapySettingsExampleView.swift
