+++ Loop/Views/Presets/Training/Components/TintedContent.swift
